Added a menu to 'yo fabric-composer'
