+++ packages/generator-composer/generators/app/index.js
@@ -28,13 +28,20 @@ module.exports = generators.Base.extend({
    * @returns {Object} List of questins to ask
    */
     prompting: function() {
+        console.log('Welcome to the Fabric Composer Skeleton Application Generator');
         const questions = [{
-            when: !this.options.appName,
-            type: 'input',
-            name: 'appName',
-            message: 'Your NPM library name:',
-            default: 'composer-sample-app',
-            store: false,
+            type: 'list',
+            name: 'generatorType',
+            message: 'Please select the type of Application:',
+            choices: [{
+                name: 'CLI Application',
+                value: 'CLI'
+            },{
+                name: 'Angular Application',
+                value: 'Angular'
+            }
+            ],
+            store: true,
             validate: function(input) {
                 if(input !== null && input !== undefined &&
           input.match(/^[\w-]+$/)) {
@@ -43,138 +50,10 @@ module.exports = generators.Base.extend({
                     return 'Name must only use lowercase letters, numbers and dashes: ^[a-z\-\d]+$';
                 }
             }
-        }, {
-            type: 'input',
-            name: 'appDescription',
-            message: 'Short description:',
-            default: 'Test Composer project',
-            store: false,
-            validate: function(input) {
-                if(input !== null && input !== undefined && input !== '') {
-                    return true;
-                } else {
-                    return 'Description cannot be null or empty.';
-                }
-            }
-        }, {
-            type: 'input',
-            name: 'authorName',
-            message: 'Author name:',
-            store: true,
-            validate: function(input) {
-                if(input !== null && input !== undefined && input !== '') {
-                    return true;
-                } else {
-                    return 'Author name cannot be null or empty.';
-                }
-            }
-        }, {
-            type: 'input',
-            name: 'authorEmail',
-            message: 'Author email:',
-            store: true,
-            validate: function(input) {
-                if(input !== null && input !== undefined && input !== '') {
-                    return true;
-                } else {
-                    return 'Author email cannot be null or empty.';
-                }
-            }
-        },{
-            type: 'input',
-            name: 'npmNetworkDependancy',
-            message: 'NPM Module name of the Business Network to connect to:',
-            default: 'digitalproperty-network',
-            store: false,
-            validate: function(input) {
-                if(input !== null && input !== undefined && input !== '') {
-                    return true;
-                } else {
-                    return 'Network cannot be null or empty.';
-                }
-            }
-        }, {
-            type: 'confirm',
-            name: 'isNpmSameAsNetworkIdentifier',
-            message: 'Is the name in NPM registry the same as the Business Network Identifier?:',
-            default: true,
-            store: false
-        }, {
-            type: 'input',
-            name: 'networkIdentifier',
-            message: 'What is the Business Network Identifier?:',
-            store: false,
-            when: function(answers) {
-                return !answers.isNpmSameAsNetworkIdentifier;
-            },
-            validate: function(input) {
-                if(input !== null && input !== undefined &&
-          input.match(/^[\/\@\w-]+$/)) {
-                    return true;
-                } else {
-                    return 'Name must only use lowercase letters, numbers and dashes: ^[a-z\-\d]+$';
-                }
-            }
-        },  {
-            type: 'input',
-            name: 'connectionProfileName',
-            message: 'What is the Connection Profile to use?',
-            default: 'defaultProfile',
-            store: false,
-            validate: function(input) {
-                if(input !== null && input !== undefined && input !== '') {
-                    return true;
-                } else {
-                    return 'Connection Profile cannot be null or empty.';
-                }
-            }
-        },{
-            type: 'input',
-            name: 'enrollmentId',
-            message: 'Enrollment id:',
-            store: true,
-            default: 'WebAppAdmin',
-            validate: function(input) {
-                if(input !== null && input !== undefined && input !== '') {
-                    return true;
-                } else {
-                    return 'Enrollment id name cannot be null or empty.';
-                }
-            }
-        }, {
-            type: 'input',
-            name: 'enrollmentSecret',
-            message: 'Enrollment Secret:',
-            default: 'DJY27pEnl16d',
-            validate: function(input) {
-                if(input !== null && input !== undefined && input !== '') {
-                    return true;
-                } else {
-                    return 'Enrollment Secret email cannot be null or empty.';
-                }
-            }
         }];
-
         return this.prompt(questions).then(answers => {
-            if (!this.options.appName) {
-                this.appName = answers.appName;
-            } else {
-                this.appName = this.options.appName;
-            }
-            if (answers.isNpmSameAsNetworkIdentifier){
-                this.networkIdentifier = answers.npmNetworkDependancy;
-            }else {
-                this.networkIdentifier = answers.networkIdentifier;
-            }
-            this.appDescription = answers.appDescription;
 
-            this.authorName = answers.authorName;
-            this.authorEmail = answers.authorEmail;
-
-            this.npmNetworkDependancy = answers.npmNetworkDependancy;
-            this.connectionProfileName = answers.connectionProfileName;
-            this.enrollmentId = answers.enrollmentId;
-            this.enrollmentSecret = answers.enrollmentSecret;
+            this.generatorType = answers.generatorType;
         });
     },
 
@@ -183,36 +62,18 @@ module.exports = generators.Base.extend({
    * Configure generator.
    */
     configuring: function() {
-        console.log('configuring: '+this.appName);
-        this.destinationRoot(this.appName);
+        if(this.generatorType == 'CLI'){
+            this.composeWith(require.resolve('../cli'));
+        }
+        else if(this.generatorType == 'Angular'){
+            this.composeWith(require.resolve('../angular'));
+        }
+        else{
+            console.log('Generator type not recognised');
+        }
     },
 
-  /**
-   * #5 in Yeoman run context.
-   * Write templates to destination.
-   */
-    writing: function() {
-        let model = this._generateTemplateModel();
-        this.fs.copyTpl(this.templatePath('**/*'), this.destinationPath(), model);
-        this.fs.move(this.destinationPath('_dot_gitignore'), this.destinationPath('.gitignore'));
-    },
 
-  /**
-   * Creates a model object passed into all templates.
-   * @return {Object} to be passed to the templates
-   */
-    _generateTemplateModel: function() {
-        return {
-            appName: this.appName,
-            appDescription: this.appDescription,
-            authorName: this.authorName,
-            authorEmail: this.authorEmail,
-            networkIdentifier: this.networkIdentifier,
-            npmNetworkDependancy: this.npmNetworkDependancy,
-            connectionProfileName: this.connectionProfileName,
-            enrollmentId: this.enrollmentId,
-            enrollmentSecret: this.enrollmentSecret
-        };
-    }
+
 
 });
