+++ packages/generator-composer/generators/cli/templates/Dockerfile
@@ -1,0 +1,2 @@
+FROM node:argon
+
