+++ packages/generator-composer/generators/cli/templates/_dot_gitignore
@@ -1,0 +1,4 @@
+node_modules/
+babel/
+coverage/
+npm-debug.log
