+++ packages/generator-composer/generators/cli/templates/config/default.json
@@ -1,0 +1,9 @@
+{
+
+  "<%= appName %>" : {
+     "participantId" : "<%= enrollmentId %>",
+     "participantPwd" :"<%= enrollmentSecret %>",
+     "businessNetworkIdentifier" : "<%= networkIdentifier %>",
+     "connectionProfile" : "<%= connectionProfileName %>"
+  }
+}
