+++ packages/generator-composer/generators/cli/templates/scripts/docker-compose.yml
@@ -1,0 +1,28 @@
+membersrvc:
+  image: hyperledger/fabric-membersrvc
+  ports:
+    - '7054:7054'
+  command: membersrvc
+vp0:
+  image: hyperledger/fabric-peer
+  ports:
+    - '7050:7050'
+    - '7051:7051'
+    - '7052:7052'
+    - '7053:7053'
+  environment:
+    - CORE_PEER_ADDRESSAUTODETECT=true
+    - CORE_VM_ENDPOINT=unix:///var/run/docker.sock
+    - CORE_LOGGING_LEVEL=DEBUG
+    - CORE_PEER_ID=vp0
+    - CORE_PEER_PKI_ECA_PADDR=membersrvc:7054
+    - CORE_PEER_PKI_TCA_PADDR=membersrvc:7054
+    - CORE_PEER_PKI_TLSCA_PADDR=membersrvc:7054
+    - CORE_SECURITY_ENABLED=true
+    - CORE_SECURITY_ENROLLID=test_vp0
+    - CORE_SECURITY_ENROLLSECRET=MwYpmSRjupbT
+  links:
+    - membersrvc
+  command: sh -c 'sleep 5; peer node start'
+  volumes:
+    - /var/run/docker.sock:/var/run/docker.sock
